# vm/core: EVM 64 bit gas instructions + optimisations