--- a/cmd/evm/main.go
+++ b/cmd/evm/main.go
@@ -29,7 +29,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/state"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethdb"
@@ -47,14 +46,6 @@ var (
 		Name:  "debug",
 		Usage: "output full trace logs",
 	}
-	ForceJitFlag = cli.BoolFlag{
-		Name:  "forcejit",
-		Usage: "forces jit compilation",
-	}
-	DisableJitFlag = cli.BoolFlag{
-		Name:  "nojit",
-		Usage: "disabled jit compilation",
-	}
 	CodeFlag = cli.StringFlag{
 		Name:  "code",
 		Usage: "EVM code",
@@ -65,18 +56,48 @@ var (
 	}
 	GasFlag = cli.StringFlag{
 		Name:  "gas",
-		Usage: "gas limit for the evm",
+		Usage: "set the gas limit for the EVM execution",
 		Value: "10000000000",
 	}
 	PriceFlag = cli.StringFlag{
 		Name:  "price",
-		Usage: "price set for the evm",
+		Usage: "set the price for the EVM execution",
 		Value: "0",
 	}
 	ValueFlag = cli.StringFlag{
 		Name:  "value",
-		Usage: "value set for the evm",
+		Usage: "set the value for the EVM execution",
 		Value: "0",
+	}
+	SenderFlag = cli.StringFlag{
+		Name:  "origin",
+		Usage: "set the origin for the EVM execution",
+		Value: "0x",
+	}
+	CoinbaseFlag = cli.StringFlag{
+		Name:  "coinbase",
+		Usage: "coinbase set for the evm",
+		Value: "0x",
+	}
+	BlockNumberFlag = cli.StringFlag{
+		Name:  "blocknumber",
+		Usage: "set the block number for the EVM execution",
+		Value: "1",
+	}
+	BlockTimeFlag = cli.StringFlag{
+		Name:  "blocktime",
+		Usage: "set the block time for the EVM execution",
+		Value: "1",
+	}
+	BlockDifficultyFlag = cli.StringFlag{
+		Name:  "difficulty",
+		Usage: "set the block difficulty for the EVM execution",
+		Value: "1",
+	}
+	GasLimitFlag = cli.StringFlag{
+		Name:  "gaslimit",
+		Usage: "sets the gas limit for the EVM execution",
+		Value: "1",
 	}
 	DumpFlag = cli.BoolFlag{
 		Name:  "dump",
@@ -105,8 +126,6 @@ func init() {
 		CreateFlag,
 		DebugFlag,
 		VerbosityFlag,
-		ForceJitFlag,
-		DisableJitFlag,
 		SysStatFlag,
 		CodeFlag,
 		CodeFileFlag,
@@ -115,6 +134,11 @@ func init() {
 		ValueFlag,
 		DumpFlag,
 		InputFlag,
+		GasLimitFlag,
+		CoinbaseFlag,
+		SenderFlag,
+		BlockNumberFlag,
+		BlockTimeFlag,
 	}
 	app.Action = run
 }
@@ -124,23 +148,49 @@ func run(ctx *cli.Context) error {
 	glog.SetV(ctx.GlobalInt(VerbosityFlag.Name))
 
 	db, _ := ethdb.NewMemDatabase()
-	statedb, _ := state.New(common.Hash{}, db)
-	sender := statedb.CreateAccount(common.StringToAddress("sender"))
 
 	logger := vm.NewStructLogger(nil)
-	vmenv := NewEnv(statedb, common.StringToAddress("evmuser"), common.Big(ctx.GlobalString(ValueFlag.Name)), common.Big(ctx.GlobalString(PriceFlag.Name)), vm.Config{
-		Debug:     ctx.GlobalBool(DebugFlag.Name),
-		ForceJit:  ctx.GlobalBool(ForceJitFlag.Name),
-		EnableJit: !ctx.GlobalBool(DisableJitFlag.Name),
-		Tracer:    logger,
-	})
 
-	tstart := time.Now()
+	db, err := ethdb.NewMemDatabase()
+	if err != nil {
+		panic(err)
+	}
+	st, err := state.New(common.Hash{}, db)
+	if err != nil {
+		panic(err)
+	}
+	backend := &core.EVMBackend{
+		GetHashFn: func(n uint64) common.Hash {
+			return common.BytesToHash(crypto.Keccak256([]byte(big.NewInt(int64(n)).String())))
+		},
+		State: st,
+	}
+	context := vm.Context{
+		CallContext: core.EVMCallContext{core.CanTransfer, core.Transfer},
+		Origin:      common.StringToAddress(ctx.GlobalString(SenderFlag.Name)),
+		Coinbase:    common.StringToAddress(ctx.GlobalString(CoinbaseFlag.Name)),
+		BlockNumber: common.String2Big(ctx.GlobalString(BlockNumberFlag.Name)),
+		Time:        common.String2Big(ctx.GlobalString(BlockTimeFlag.Name)),
+		Difficulty:  common.String2Big(ctx.GlobalString(BlockDifficultyFlag.Name)),
+		GasLimit:    common.String2Big(ctx.GlobalString(GasLimitFlag.Name)),
+		GasPrice:    common.String2Big(ctx.GlobalString(PriceFlag.Name)),
+	}
+
+	vmenv := vm.NewEnvironment(
+		context,
+		backend,
+		params.TestChainConfig,
+		vm.Config{
+			Debug:  ctx.GlobalBool(DebugFlag.Name),
+			Tracer: logger,
+		},
+	)
 
 	var (
-		code []byte
-		ret  []byte
-		err  error
+		code   []byte
+		ret    []byte
+		tstart = time.Now()
+		sender = st.CreateAccount(context.Origin)
 	)
 
 	if ctx.GlobalString(CodeFlag.Name) != "" {
@@ -174,8 +224,7 @@ func run(ctx *cli.Context) error {
 			common.Big(ctx.GlobalString(ValueFlag.Name)),
 		)
 	} else {
-		receiver := statedb.CreateAccount(common.StringToAddress("receiver"))
-
+		receiver := st.CreateAccount(common.StringToAddress("receiver"))
 		receiver.SetCode(crypto.Keccak256Hash(code), code)
 		ret, err = vmenv.Call(
 			sender,
@@ -188,8 +237,7 @@ func run(ctx *cli.Context) error {
 	vmdone := time.Since(tstart)
 
 	if ctx.GlobalBool(DumpFlag.Name) {
-		statedb.Commit(true)
-		fmt.Println(string(statedb.Dump()))
+		fmt.Println(string(st.Dump()))
 	}
 	vm.StdErrFormat(logger.StructLogs())
 
@@ -219,90 +267,4 @@ func main() {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
-}
-
-type VMEnv struct {
-	state *state.StateDB
-	block *types.Block
-
-	transactor *common.Address
-	value      *big.Int
-
-	depth      int
-	Gas, price *big.Int
-	time       *big.Int
-	logs       []vm.StructLog
-
-	evm *vm.EVM
-}
-
-func NewEnv(state *state.StateDB, transactor common.Address, value, price *big.Int, cfg vm.Config) *VMEnv {
-	env := &VMEnv{
-		state:      state,
-		transactor: &transactor,
-		value:      value,
-		price:      price,
-		time:       big.NewInt(time.Now().Unix()),
-	}
-
-	env.evm = vm.New(env, cfg)
-	return env
-}
-
-// ruleSet implements vm.ChainConfig and will always default to the homestead rule set.
-type ruleSet struct{}
-
-func (ruleSet) IsHomestead(*big.Int) bool { return true }
-func (ruleSet) GasTable(*big.Int) params.GasTable {
-	return params.GasTableHomesteadGasRepriceFork
-}
-
-func (self *VMEnv) ChainConfig() *params.ChainConfig { return params.TestChainConfig }
-func (self *VMEnv) Vm() vm.Vm                        { return self.evm }
-func (self *VMEnv) Db() vm.Database                  { return self.state }
-func (self *VMEnv) SnapshotDatabase() int            { return self.state.Snapshot() }
-func (self *VMEnv) RevertToSnapshot(snap int)        { self.state.RevertToSnapshot(snap) }
-func (self *VMEnv) Origin() common.Address           { return *self.transactor }
-func (self *VMEnv) BlockNumber() *big.Int            { return common.Big0 }
-func (self *VMEnv) Coinbase() common.Address         { return *self.transactor }
-func (self *VMEnv) Time() *big.Int                   { return self.time }
-func (self *VMEnv) Difficulty() *big.Int             { return common.Big1 }
-func (self *VMEnv) BlockHash() []byte                { return make([]byte, 32) }
-func (self *VMEnv) Value() *big.Int                  { return self.value }
-func (self *VMEnv) GasLimit() *big.Int               { return big.NewInt(1000000000) }
-func (self *VMEnv) GasPrice() *big.Int               { return big.NewInt(1000000000) }
-func (self *VMEnv) VmType() vm.Type                  { return vm.StdVmTy }
-func (self *VMEnv) Depth() int                       { return 0 }
-func (self *VMEnv) SetDepth(i int)                   { self.depth = i }
-func (self *VMEnv) GetHash(n uint64) common.Hash {
-	if self.block.Number().Cmp(big.NewInt(int64(n))) == 0 {
-		return self.block.Hash()
-	}
-	return common.Hash{}
-}
-func (self *VMEnv) AddLog(log *vm.Log) {
-	self.state.AddLog(log)
-}
-func (self *VMEnv) CanTransfer(from common.Address, balance *big.Int) bool {
-	return self.state.GetBalance(from).Cmp(balance) >= 0
-}
-func (self *VMEnv) Transfer(from, to vm.Account, amount *big.Int) {
-	core.Transfer(from, to, amount)
-}
-
-func (self *VMEnv) Call(caller vm.ContractRef, addr common.Address, data []byte, gas, value *big.Int) ([]byte, error) {
-	self.Gas = gas
-	return core.Call(self, caller, addr, data, gas, value)
-}
-
-func (self *VMEnv) CallCode(caller vm.ContractRef, addr common.Address, data []byte, gas, value *big.Int) ([]byte, error) {
-	return core.CallCode(self, caller, addr, data, gas, value)
-}
-
-func (self *VMEnv) DelegateCall(caller vm.ContractRef, addr common.Address, data []byte, gas *big.Int) ([]byte, error) {
-	return core.DelegateCall(self, caller, addr, data, gas)
-}
-
-func (self *VMEnv) Create(caller vm.ContractRef, data []byte, gas, value *big.Int) ([]byte, common.Address, error) {
-	return core.Create(self, caller, data, gas, value)
 }