**`common/math, core/vm, core/state: EVM gas calculations to 64bit`**:

This changes the internals of the EVM-JIT by dropping support for blocks (and therefor also transactions) that can exceed a 64^2-1 gas-limit. This means that the EVM-JIT can drop support for abritrary precision (*big.Int) and move to 64bit based gas calculations.

For obvious reasons integer overflows have added to make sure that gas calculations such as gas for memory doesn't overflow 64bit. If the overflow check fails it throws an out-of-gas error and will halt the execution of the code.

**`core/vm: removed arbitrary precision VM + multithreaded JIT status`**

Removed the arbitrary precision VM in favour of the optimised JIT VM. This enables the new VM for all calls and creates and removed the need for the --jitvm and --forcejit flags.

This fixes a minor issue when the jit is used in a multithreaded environment it would ignore the compilation status of the JIT and default to the byte-code VM. Instead it now waits for the compilation to finish and run the JIT if the execution was successful.

**`core/vm, tests: refactored native contracts`**

Refactored native contract such that native contract can be easily extended using a struct that implements the vm.NativeContract interface.

**`core, tests, miner, cmd: improved EVM environment`**

Reduced the overhead caused by all the diffirent implementations of the vm.Environment interface. Where previously the Environment was an interface it has now become an implementation it self using several new easy to implement interfaces such as the Backend that takes care of the
state and CallContext which takes care of the EVM calling conventions.

Most of these interfaces have been implemented in the core package with the exception of the test suit which uses a combination of the core package implementations and a mix of its own.

This will reduce the amount of extra overhead when changing an interface, such as adding a new method or changing a method signature.
#### [HISTORY] Original description

This PR changes some of the internals of the EVM, namely it changes how gas is accounted for. It changes most the gas arithmetic from `*big.Int` to `uint64` where possible and constrains the memory to `uint64` instead of a possible, impractical unlimited size.

Changing the gas counter from `big.Int` to `uint64` [reduces](https://gist.github.com/obscuren/6bf7d36c59e3a6f955dc744004ed1d1f) the heap allocation significantly and reduces the arithmetic complexity of the programs.

This changes is only done to the new JIT VM and will not be done for the byte code EVM (subject to change).
### Tasks
- [x] Bench `*big.Int` methods `BitLen` and `len(Bytes())` (when constraining `BitLen() > 64` or `len(Bytes()) > 8`
- [x] Convert gas calculations from `*big.Int` to `uint64`
- [x] Convert `segments` from `*big.Int` to `uint64`
  - [x] `pushSeg`
  - [x] `jumpSeg`
- [ ] Optimise overflow functions (bit check & contstraint `2^63-1`) or cpu [overflow flag](https://en.wikipedia.org/wiki/Overflow_flag)
- [x] `contract.ReturnGas` to `uint64`
- [x] Remove `contract.ReturnGas` `Price` argument
### TODO
- [ ] Implement `GetRefund` tests (block/vm/tests does not cover refund)