@obscuren 2016-05-17
Bench results of `len(num.Bytes()) > 8` and `num.BitLen() > 64`:

```
BenchmarkBitlenLarge-4  	200000000	         7.00 ns/op
BenchmarkBitlenSmall-4  	200000000	         7.06 ns/op
BenchmarkBytesLenLarge-4	20000000	        70.5 ns/op
BenchmarkBytesLenSmall-4	20000000	        60.9 ns/op
```

`BitLen` is clearly the winner.