@fjl 2016-07-12
@obscuren Are you going to get back to this? I would love to see this happen for 1.5.