@obscuren 2016-07-14
Benchmarks for EVM bytecode vs 64bit JIT:

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkStateCall1024-4          25955         24733         -4.71%
BenchmarkVmAckermann32Tests-4     9252061       1915683       -79.29%
BenchmarkVmFibonacci16Tests-4     55585810      11739691      -78.88%
```