@GitCop 2016-07-19
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: cf3d0381592a873ade9f4ccf8aebc29706cf2a9b
  * Commits must be prefixed with the package(s) they modify

 * Commit: b1f22f462520375e0e6c2e314ce06187b2e252d3
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com