@GitCop 2016-07-19
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: 8867b37400410e5af4a7950fa9e81842b3be47ab
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com