@obscuren 2016-08-19
@fjl maybe this will be ready for 1.5 after all.