@GitCop 2016-09-22
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: c380cd4e2d6f31c95c125008be21674a3e70a0b5
  * Commits must be prefixed with the package(s) they modify

 * Commit: 474a20040bdebefa4082ff3b3bd1dea7c9678878
  * Commits must be prefixed with the package(s) they modify

 * Commit: 7f818b1445990c0370709e431fdfaf24c3af1374
  * Commits must be prefixed with the package(s) they modify

 * Commit: c82c5b022f0ce96e7c80599307872609e6527eea
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com