@GitCop 2016-10-09
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: fe44f6cbf468665c9aad6ff6822c0549bfed0357
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com