@GitCop 2016-10-10
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: d7107154791e348424c1c57119da551e20301aa6
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com