@GitCop 2016-10-27
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: 0169435782144cbc5680908f722ca154f6587ee0
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com