@GitCop 2016-11-03
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: e65d1766a30460a7b9393c2f2ffbd9cd8a56abd4
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com